+++ examples/concurrency/threads.rb
@@ -1,9 +1,9 @@
 require File.expand_path('../../helper.rb', __FILE__)
 
-locales = AmazonProduct::Locale::LOCALES
+locales = Vacuum::Locale::LOCALES
 
 locales.each do |locale|
-  AmazonProduct[locale].configure do |c|
+  Vacuum[locale].configure do |c|
     c.key    = AMAZON_KEY
     c.secret = AMAZON_SECRET
     c.tag    = AMAZON_ASSOCIATE_TAG
@@ -12,7 +12,7 @@ end
 
 threads = locales.map do |locale|
   Thread.new do
-    req = AmazonProduct[locale]
+    req = Vacuum[locale]
     Thread.current[:resp] = req.find('0143105825')
   end
 end