+++ lib/vacuum/cart_operations.rb
@@ -1,10 +1,10 @@
-module AmazonProduct
+module Vacuum
   # Cart operations
   module CartOperations
     # Creates a cart
     #
     # @param [Hash] params
-    # @return [AmazonProduct::Cart] a cart
+    # @return [Vacuum::Cart] a cart
     def create_cart(params)
       Cart.new(self, params)
     end