+++ spec/spec_helper.rb
@@ -7,7 +7,7 @@ begin
 rescue LoadError
 end
 
-require File.expand_path('../../lib/amazon_product', __FILE__)
+require File.expand_path('../../lib/vacuum', __FILE__)
 
 RSpec.configure do |c|
   c.treat_symbols_as_metadata_keys_with_true_values = true